Refactoring: avoid unnecessary casts and (un)boxing

+++ client/src/main/java/org/evosuite/testcase/TestChromosome.java
@@ -250,7 +250,7 @@ public class TestChromosome extends ExecutableChromosome {
 		if (lastExecutionResult != null && !isChanged()) {
 			Integer lastPos = lastExecutionResult.getFirstPositionOfThrownException();
 			if (lastPos != null)
-				lastPosition = lastPos.intValue();
+				lastPosition = lastPos;
 		}
 
 		for (TestMutationHistoryEntry mutation : mutationHistory) {
